requirements.txt
amqp==2.5.2
asgiref==3.2.5
asn1crypto==1.3.0
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coincurve==13.0.0
cytoolz==0.10.1
Django==2.2.7
django-celery-results==1.2.1
eciespy==0.3.5
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
hexbytes==0.2.0
idna==2.9
ipfshttpclient==0.4.12
jsonschema==3.2.0
keyring==21.2.0
kombu==4.6.8
lru-dict==1.1.6
multiaddr==0.0.9
netaddr==0.7.19
parsimonious==0.8.1
protobuf==3.11.3
pycparser==2.20
pycryptodome==3.9.7
PyMySQL==0.9.3
pypiwin32==223
pyrsistent==0.15.7
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
redis==3.4.1
requests==2.23.0
rlp==1.2.0
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
toolz==0.10.0
urllib3==1.25.8
varint==1.0.2
vine==1.3.0
web3==5.7.0
websockets==8.1
